refactor: 채팅방 목록 조회 훅(useChatRooms, useMyChatRooms) 관심사 분리 및 서비스 함수 분리
